+++ elements/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+## Changed
+
+- [pos-image](../docs/elements/components/pos-image): Added alt attribute
+- [pos-picture](../docs/elements/components/pos-picture): Adds resource label as alt attribute to image
+
 ## 0.7.0
 
 ## Added